+++ .gitignore
@@ -2,4 +2,5 @@
 .eslintcache
 /node_modules
 npm-debug.log
-lib
+yarn-error.log
+dist
