+++ package.json
@@ -4,30 +4,29 @@
   "license": "MIT",
   "version": "3.2.0",
   "files": [
-    "lib"
+    "dist"
   ],
-  "engines": {
-    "node": ">=6.5.0",
-    "npm": ">=3.0.0"
-  },
-  "main": "lib/index.js",
+  "main": "dist/index.js",
   "devDependencies": {
-    "babel-cli": "^6.26.0",
-    "babel-eslint": "^8.0.1",
-    "babel-preset-es2015": "^6.24.1",
-    "babel-preset-stage-0": "^6.24.1",
-    "eslint": "^4.7.2",
-    "eslint-config-standard": "^10.2.1",
-    "eslint-plugin-import": "^2.7.0",
-    "eslint-plugin-node": "^5.1.1",
-    "eslint-plugin-promise": "^3.5.0",
-    "eslint-plugin-standard": "^3.0.1"
+    "@typescript-eslint/eslint-plugin": "^1.12.0",
+    "@typescript-eslint/parser": "^1.12.0",
+    "eslint": "^6.0.1",
+    "eslint-config-prettier": "^6.0.0",
+    "eslint-config-standard": "^13.0.1",
+    "prettier": "^1.18.2",
+    "rimraf": "^2.6.3",
+    "typescript": "^3.5.3"
+  },
+  "prettier": {
+    "semi": false,
+    "singleQuote": true,
+    "trailingComma": "all"
   },
   "scripts": {
-    "test": "eslint --cache --ext .js,.jsx src",
-    "prebuild": "rm -rf ./lib && npm test",
-    "build": "babel --quiet ./src -d lib",
+    "lint": "eslint --cache --ext .ts,.tsx src",
+    "prebuild": "rimraf dist",
+    "build": "tsc -p tsconfig.json",
     "postversion": "git push --tags origin HEAD",
-    "prepublish": "npm run build"
+    "prepublishOnly": "yarn lint && yarn build"
   }
 }
