--- src/index.js
@@ -1,4 +1,0 @@
-export ColorPalette from './ColorPalette'
-export alpha from './alpha'
-export hexToRgb from './hexToRgb'
-export rgbToHex from './rgbToHex'
