+++ src/index.ts
@@ -1,0 +1,4 @@
+export { default as ColorPalette } from './ColorPalette'
+export { default as alpha } from './alpha'
+export { default as hexToRgb } from './hexToRgb'
+export { default as rgbToHex } from './rgbToHex'
