+++ src/rgbToHex.ts
@@ -1,4 +1,8 @@
-export default function rgbToHex(red, green, blue) {
+export default function rgbToHex(
+  red: number,
+  green: number,
+  blue: number,
+): string {
   const rgb = blue | (green << 8) | (red << 16)
   return '#' + (0x1000000 + rgb).toString(16).slice(1)
 }
